+++ packages/rx_bloc_cli/lib/src/models/generator_arguments_provider.dart
@@ -120,6 +120,10 @@ class GeneratorArgumentsProvider {
     final changeLanguageEnabled =
         _reader.read<bool>(CreateCommandArguments.changeLanguage);
 
+    // Remote translations
+    final remoteTranslationsEnabled =
+        _reader.read<bool>(CreateCommandArguments.remoteTranslations);
+
     // Analytics, Push Notifications, Firebase
     final analyticsEnabled =
         _reader.read<bool>(CreateCommandArguments.analytics);
@@ -161,6 +165,7 @@ class GeneratorArgumentsProvider {
 
     return FeatureConfiguration(
       changeLanguageEnabled: changeLanguageEnabled,
+      remoteTranslationsEnabled: remoteTranslationsEnabled,
       analyticsEnabled: analyticsEnabled,
       pushNotificationsEnabled: pushNotificationsEnabled,
       realtimeCommunicationEnabled: realtimeCommunicationEnabled,