+++ packages/rx_bloc_cli/mason_templates/bricks/rx_bloc_base/__brick__/test/helpers/golden_helper.dart
@@ -5,8 +5,9 @@ import 'package:flutter_test/flutter_test.dart';
 
 import 'package:{{project_name}}/assets.dart';
 import 'package:{{project_name}}/base/theme/design_system.dart';
-import 'package:{{project_name}}/base/theme/{{project_name}}_theme.dart';
-import 'package:{{project_name}}/l10n/{{project_name}}_app_i18n.dart';
+
+import 'package:{{project_name}}/base/theme/{{project_name}}_theme.dart';{{#enable_remote_translations}}
+import 'package:{{project_name}}/l10n/{{project_name}}_app_i18n.dart';{{/enable_remote_translations}}
 
 import 'enums/app_themes.dart';
 import 'enums/golden_alignment.dart';
@@ -186,8 +187,10 @@ Future<void> pumpDeviceBuilderWithLocalizationsAndTheme(
     pumpScenarioBuilderWithMaterialApp(
         tester,
         widget,
-        localizations: const [
-          AppI18n.delegate,
+        localizations: const [ {{#enable_remote_translations}}
+          AppI18n.delegate,{{/enable_remote_translations}}{{^enable_remote_translations}}
+          I18n.delegate,{{/enable_remote_translations}}
+          ...GlobalMaterialLocalizations.delegates,
           GlobalMaterialLocalizations.delegate,
         ],
         localeOverrides: I18n.supportedLocales,